% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUnderstanding.R
\name{AbsoluteValues}
\alias{AbsoluteValues}
\title{Display the target binary variable}
\usage{
AbsoluteValues(DataFrame, targetVariable)
}
\arguments{
\item{DataFrame}{data.frame.}

\item{targetVariable.}{character. Name of (target) variable to display}
}
\value{

}
\description{
Display a binary (target) variable in ggplot2.
}
\examples{
## Call
AbsoluteValues(DataFrame = mtcars, targetVariable = "am")
}
