% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUnderstanding.R
\name{AverageByCategoricalVariable}
\alias{AverageByCategoricalVariable}
\title{Average of one or more numerical features by a categorical (target) variable
and the single `CategoricalVariable`}
\usage{
AverageByCategoricalVariable(
  NumericalVariables,
  CategoricalVariable,
  DataFrame,
  dropNA = FALSE
)
}
\arguments{
\item{NumericalVariables}{character. Vector of names of numerical variables.}

\item{CategoricalVariable}{character. Name of single categorical (target) variable.}

\item{DataFrame}{data.frame. Input data.frame. Has to contain the `NumericalVariables`}

\item{dropNA}{boolean. Drop NA values prior to analysis in numerical variable? Default is `FALSE`}
}
\value{

}
\description{
Average of one or more numerical features by a categorical (target) variable
and the single `CategoricalVariable`
}
\examples{
# a few numerical features
AverageByCategoricalVariable(NumericalVariables = c("Sepal.Length", "Sepal.Width"),
                             CategoricalVariable = c("Species"),
                             DataFrame = iris,
                             dropNA = TRUE)
# a single numerical feature
AverageByCategoricalVariable(NumericalVariables = c("mpg"),
                             CategoricalVariable = c("am"),
                             DataFrame = mtcars,
                             dropNA = TRUE)
# Example with NA values
AverageByCategoricalVariable(NumericalVariables = c("Ozone"),
                             CategoricalVariable = c("Month"),
                             DataFrame = airquality,
                             dropNA = FALSE)
# Drop NA values
AverageByCategoricalVariable(NumericalVariables = c("Ozone"),
                             CategoricalVariable = c("Month"),
                             DataFrame = airquality,
                             dropNA = TRUE)
}
