% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnderDevelopment.R
\name{BerechnungTestStatistiken}
\alias{BerechnungTestStatistiken}
\title{Berechnung einiger zentraler Evaluationsparameter: Vorhersage Metrischer Variablen}
\usage{
BerechnungTestStatistiken(Labels, Predictions)
}
\arguments{
\item{Labels}{numeric. Vektor der metrisch skalierten Zielvariablen. NA-Werte sind aktuell nicht erlaubt und 
müssen noch manuell entfernt werden}

\item{Predictions}{numeric. Vektor der auf Basis des Machine Learning Models vorhergesagten 
metrisch-skalierten Variablen}
}
\value{
character. Named Vektor mit zentralen Evaluationsparametern: MAE, MSE, RMSE und R2
}
\description{
Berechnung einiger zentraler Evaluationsparameter: Vorhersage Metrischer Variablen
}
\examples{
BerechnungTestStatistiken(Labels = c(0.00, 1.50, 3.50, 6.30, 1.50, 0.00, 3.00, 4.16),
                          Predictions = c(0.1653140, 2.8294117, 3.5218980, 5.0807042, 2.9414713, 
                                         -0.1564419, 3.0363548, 2.4665890))
}
