% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUnderstanding.R
\name{ChangeEncodingInAnObject}
\alias{ChangeEncodingInAnObject}
\title{Change Encoding in an existing R data.frame}
\usage{
ChangeEncodingInAnObject(
  InputDataFrame,
  TargetEncoding = "ansi",
  ConvertFactorIntoCharacter = TRUE
)
}
\arguments{
\item{InputDataFrame}{data.frame. Input data.frame}

\item{TargetEncoding}{character. Target encoding. default is `ansi`. Check `?Èncoding` function for more information
about available encodings.}

\item{ConvertFactorIntoCharacter}{logical. Should factor variables first converted in character variable. If set to `FALSE`,
this will not be done and (!) labels of factor variables will not be changed into the target encoding. default is `TRUE`}
}
\value{
data.frame. data.frame with target encoding.
}
\description{
Change Encoding in an existing R data.frame
}
\examples{
# Change into ANSI
DataFrame <- data.frame(id = 1:3, Position = c("Mähen", "Flüssigdünger ausbringen", "Mähen diagonal"))
DataFrame
# Change into UTF-8
DataFrameUTF8 <- ChangeEncodingInAnObject(InputDataFrame = DataFrameANSI, TargetEncoding = "UTF-8")
DataFrameUTF8
# Change back
ChangeEncodingInAnObject(InputDataFrame = DataFrameUTF8, TargetEncoding = "ANSI")
}
