% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUnderstanding.R
\name{crossTableRowPercentages}
\alias{crossTableRowPercentages}
\title{Cross table with row precentages}
\usage{
crossTableRowPercentages(column, row)
}
\arguments{
\item{column}{vector. Column Vector: Should be target variable}

\item{row}{vector. Row Vector}
}
\value{
table. Crosstable
}
\description{
Cross table with row precentages
}
\examples{
# mtcars
crossTableRowPercentages(column = mtcars$am, row = mtcars$cyl)
}
