% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUnderstanding.R
\name{DisplayCoefficients}
\alias{DisplayCoefficients}
\title{Display Coefficients}
\usage{
DisplayCoefficients(CoefficientVectorOrlmModel, DisplayIntercept = F)
}
\arguments{
\item{CoefficientVectorOrlmModel}{either a named vectoror a lm-model result.}

\item{DisplayIntercept}{boolean. If TRUE, everytrhing that sounds like `intercept`
will be deleted out of the dataset}
}
\value{
data.frame. DataFrame of coefficients with three variables: variablenames,
Coefficients, PositiveDirection
}
\description{
Displays either coefficients from a named vector or from
a lm-model.
}
\details{
If the character vector has no coefficient names, a character
vector in form v1...vk (k: number of coefficients in vector) will be built.
}
\examples{
## Mode 1: named vector
# Generate some data
Coefficients <- c(0.5, 0.7, -0.3)
names(Coefficients) <- c("Klarheit", "Freundlichkeit", "Dauer")
## Call function
DisplayCoefficients(CoefficientVectorOrlmModel = Coefficients,
                    DisplayIntercept = F)
## Mode 2: lm model output
# Generate a lm-object:
mtcarsLmComplete <- lm(mpg ~ ., data = mtcars)
## Call function
DisplayCoefficients(CoefficientVectorOrlmModel = mtcarsLmComplete,
                    DisplayIntercept = T)
}
