% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUnderstanding.R
\name{FeatureByNPS}
\alias{FeatureByNPS}
\title{Displays a feature by a Net Promoter Score (NPS) variable}
\usage{
FeatureByNPS(DataFrame, Feature, NPSVariable)
}
\arguments{
\item{DataFrame}{data.frame. data.frame under investigation.}

\item{Feature}{vector. Feature variable.}

\item{NPSVariable}{factor. Net Promoter Score with categories.}
}
\value{

}
\description{
A special modification of function FeatureByTargetVariable to incorporate the commonly
used Net Promoter Score Variable with three levels (Detractor, Neutral and Promoter)
}
\details{
: Please make sure that the colors match the categories - Currently, alphabetical order of categories
will be used.
}
\examples{
# generate a small dataset
NPSdata <- data.frame(Feat = sample(x = 1:4, size = 20, replace = T),
                      NPS = sample(x = c("Detractor", "Neutral", "Promoter"), size = 20, replace = T))
FeatureByNPS(DataFrame = NPSdata, Feature = "Feat", NPSVariable = "NPS")
}
