% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUnderstanding.R
\name{FeatureByTargetVariable}
\alias{FeatureByTargetVariable}
\title{Displays a feature variable by a binary target variable}
\usage{
FeatureByTargetVariable(DataFrame, Feature, TargetVariable)
}
\arguments{
\item{DataFrame}{data.frame. data.frame under investigation.}

\item{Feature}{vector. Feature variable.}

\item{TargetVariable}{factor. Binary target variable.}
}
\value{

}
\description{
Displays a feature variable by a binary target variable
}
\examples{
## mtcars
FeatureByTargetVariable(DataFrame = mtcars, Feature = "gear", TargetVariable = "am")
## Selected variables in data.frame:
Variables <- c("cyl", "gear", "vs")
for (i in Variables) {
    print(FeatureByTargetVariable(DataFrame = mtcars, Feature = i, TargetVariable = "am"))
}
}
