% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUnderstanding.R
\name{freqTable}
\alias{freqTable}
\title{Frequency table with number of cases and relative number of cases}
\usage{
freqTable(x)
}
\arguments{
\item{x}{vector. Should be a vector.}
}
\value{
data.frame. absolute and relative values
}
\description{
Provides a table with absolute and relative number of cases.
The output table can be easily used in Markdown with kable().
}
\examples{
## iris
freqTable(x = iris$Species)
## mtcars
# automatic or manual?
freqTable(x = mtcars$am)
# number of gears
freqTable(x = mtcars$gear)
}
