% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUnderstanding.R
\name{FreqTableMissingValues}
\alias{FreqTableMissingValues}
\title{Count the number of missing values in each variable in a data.frame}
\usage{
FreqTableMissingValues(DataFrame, absoluteValues = TRUE, chart = FALSE)
}
\arguments{
\item{DataFrame}{data.frame. Contains the variables which should be
investigated.}

\item{absoluteValues.}{boolean. If TRUE, absolute values will be printed out,
if FALSE, relative values. Default are absolute values.}
}
\value{
data.frame. Returns a data.frame with variable names, number of valid
cases and number of missing cases.
}
\description{
Counts the number of missing values of each variable in a
data.frame and returns either absolute or relative values
(argument `absoluteValues`)
}
\examples{
## dataset airquality contains NAs
airquality[1:8,]
# absolute values
FreqTableMissingValues(DataFrame = airquality, absoluteValues = TRUE, chart = FALSE)
# relative values with chart
FreqTableMissingValues(DataFrame = airquality, absoluteValues = FALSE, chart = TRUE)
}
