% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUnderstanding.R
\name{LoadAndOrInstallAPackage}
\alias{LoadAndOrInstallAPackage}
\title{Load and/or install a package}
\usage{
LoadAndOrInstallAPackage(PackageName)
}
\arguments{
\item{PackageName}{character. Name of the package to load / install and load.}
}
\value{
nothing.
}
\description{
Load and/or install a package
}
\examples{
LoadAndOrInstallAPackage(PackageName = "ggplot2")
LoadAndOrInstallAPackage(PackageName = "effsize")
}
