% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUnderstanding.R
\name{numberedVariablesList}
\alias{numberedVariablesList}
\title{PrintDisplay numbered list of variables in a dataset}
\usage{
numberedVariablesList(df)
}
\arguments{
\item{data.frame.}{Input object: data.frame.}
}
\value{
data.frame. data.frame with row numbers corresponding to the position of the
variables in the dataset and variable names.
}
\description{
Often it is really useful to have a numbered list of variables
in a dataset, especially if the dataset is new. The function will provide such
a list
}
\examples{
# iris
numberedVariablesList(df = iris)
# mtcars
numberedVariablesList(df = mtcars)
}
