% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUnderstanding.R
\name{NumericalFeatureByTargetVariable}
\alias{NumericalFeatureByTargetVariable}
\title{Numerical Feature by categorical target variable}
\usage{
NumericalFeatureByTargetVariable(DataFrame, NumericalFeature, TargetVariable)
}
\arguments{
\item{NumericalFeature}{character. Numerical feature}

\item{TargetVariable}{character. Categorical target variable.}

\item{DataFrame.}{data.frame. data.frame under investigation.}
}
\value{
chart.
}
\description{
Numerical Feature by categorical target variable
}
\examples{
NumericalFeatureByTargetVariable(DataFrame = iris,
                                 NumericalFeature = "Sepal.Length",
                                 TargetVariable = "Species")
}
