% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUnderstanding.R
\name{RenameVariable}
\alias{RenameVariable}
\title{Renames a single variable in a data.frame}
\usage{
RenameVariable(InputDataFrame, ExistingVariableName, NewVariableName)
}
\arguments{
\item{InputDataFrame}{data.frame. Input data.frame in which the column name of a **single** variable should be 
changed.}

\item{ExistingVariableName}{character. Existing column name, which should be changed}

\item{NewVariableName}{character. New column name.}
}
\value{
data.frame. Renamed - (!)complete - data.frame.
}
\description{
Renames a single variable in a data.frame
}
\examples{
RenameVariable(InputDataFrame = iris, ExistingVariableName = "Sepal.Length", NewVariableName = "SepalLength")
}
