% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnderDevelopment.R
\name{ResiduenBereinigen}
\alias{ResiduenBereinigen}
\title{Entfernt standardisierte Residuen eines lm-Objekts außerhalb des angegebenen Schwellenwertes}
\usage{
ResiduenBereinigen(lmObjekt, SchwellenwertAusschlussStandResiduen)
}
\arguments{
\item{lmObjekt}{lm. lm-Objekt als Ergebnis der lm()-Funktion.}

\item{SchwellenwertAusschlussStandResiduen}{numerisch. Schwellenwert in z-Einheiten. default ist 2.}
}
\value{
logical. Logischer Vektor, TRUE stellt einen Wert außerhalb des Schwellenwertes vor. Kann als Filter zur 
Bereinigung des  data.frame genutzt werden.
}
\description{
Die Funktion hat als Parameter ein lm-Objekt und einen Schwellenwert als z-Wert
Zuerst werden die vorhandenen Residuen des lm-Objekts z-standardisiert grafisch darstellt, dann werden
Fälle außerhalb des angegebenen Schwellenwertes identifiziert und eine um diese Fälle bereinigte Grafik
angezeigt. Es wird ein Filter mit den außerhalb liegenden Fällen als TRUE zurückgegeben.
}
\examples{
mtcarsLMObjekt <- lm(mpg ~ ., data = mtcars)
ResiduenBereinigen(lmObjekt = mtcarsLMObjekt, SchwellenwertAusschlussStandResiduen = 2)
}
