% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUnderstanding.R
\name{skewness}
\alias{skewness}
\title{Skewness}
\usage{
skewness(InputVariable, na.rm = FALSE)
}
\arguments{
\item{InputVariable}{numerical or logical. Inputvariable has to be a numerical or logical vector.
Other inputs are not allowed. NA values are allowed and can be removed using argument `na.rm`}

\item{na.rm}{logical- Logical value if NA values should be removed. default is `FALSE`, NAs will not be removed and result is NA.
Set to `TRUE` to remove NA values.}
}
\value{
numerical. Returns a single value representing the skewness.
}
\description{
Skewness
}
\examples{
skewness(InputVariable = rnorm(1E7)) # result should be approx. 0
skewness(InputVariable = rlnorm(1E7)) # result should be > 5
}
