% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnderDevelopment.R
\name{SplitUndEntfernenVonKonstanten}
\alias{SplitUndEntfernenVonKonstanten}
\title{Split data.frame in Trainings- und Testdatensatz und Bereinigung}
\usage{
SplitUndEntfernenVonKonstanten(DataFrame, SplitRatio = 0.7, NAEntfernen = TRUE)
}
\arguments{
\item{DataFrame}{data.frame. data.frame, der in einen Trainings- und Testdatensatz zerlegt werden soll.}

\item{SplitRatio}{double. Anteil an Fällen im Trainingsdatensatz - **in Punkten anzugeben!** - default: `.7` (70%)}

\item{NAEntfernen}{logicial. Sollen Zeilen mit fehlenden Werten entfernt wernden? Es findet ein sog. 
Fallweiser (listwise) Ausschluss statt, d.h. der ganze Fall wird entfernt. default = `TRUE`.}
}
\value{
list. Liste mit den beiden data.frames: Trainings- und Test data.frames.
}
\description{
Splittet einen data.frame nach der Split Ratio (default .7) in einen Trainings- und Testdatensatz und 
entfernt Variablen ohne Varianz in beiden Splits sowie Zeilen mit NAs
}
\examples{
TrainingUndTestDataFrame <- SplitUndEntfernenVonKonstanten(DataFrame = AnalyseDatenSatz, SplitRatio = .7, NAEntfernen = TRUE)
sapply(X = TrainingUndTestDataFrame, dim)
}
