% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUnderstanding.R
\name{VariableMetaData}
\alias{VariableMetaData}
\title{Displays Metadata of variables in a data.frame}
\usage{
VariableMetaData(DataFrame)
}
\arguments{
\item{DataFrame}{data.frame. data to be analyzed. data.frame required.}
}
\value{
data.frame. data.frame with metadata.
}
\description{
Displays Metadata of variables in a data.frame
}
\details{
Returns four columns in a data.frame:
1: VarNo:         Running Number
2: VarNames:      Variable Name
3: StorageType:   Atomic class
4: UniqueValues:  Number of unique categories
}
\examples{
## famous iris data set
VariableMetaData(DataFrame = iris)
## mtcars data set
VariableMetaData(DataFrame = mtcars)
}
