% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUnderstanding.R
\name{freqTable}
\alias{freqTable}
\title{Frequency table with number of cases and relative number of cases}
\usage{
freqTable(x, printNA = TRUE)
}
\arguments{
\item{x}{vector. Should be a vector.}

\item{printNA}{boolean. Print NA values? Default is TRUE}
}
\value{
data.frame. absolute and relative values
}
\description{
Provides a table with absolute and relative number of cases.
The output table can be easily used in Markdown with kable().
}
\details{
Per default NA values will be displayed, no matter if there are any.
Set `printNA` argument to FALSE to surpress NA values. Please note: If there are
any NA values in x (variable), a warning will be issued!
}
\examples{
## iris
freqTable(x = iris$Species, printNA = TRUE)
## mtcars
# automatic or manual?
freqTable(x = mtcars$am, printNA = FALSE)
# number of gears
freqTable(x = mtcars$gear, printNA = FALSE)
# NA values in variable
freqTable(x = airquality$Ozone, printNA = TRUE)
freqTable(x = airquality$Ozone, printNA = FALSE)
}
